backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-editable --format=requirements.txt
.
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
    #   starlette
asgiref==3.11.1
    # via opentelemetry-instrumentation-asgi
asyncpg==0.31.0
    # via databricks-apps-fastapi-starter
cachetools==7.0.5
    # via mlflow-skinny
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   mlflow-skinny
    #   uvicorn
cloudpickle==3.1.2
    # via mlflow-skinny
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
cryptography==46.0.5
    # via google-auth
databricks-sdk==0.100.0
    # via
    #   databricks-apps-fastapi-starter
    #   mlflow-skinny
databricks-sql-connector==4.2.5
    # via databricks-apps-fastapi-starter
databricks-vectorsearch==0.66
    # via databricks-apps-fastapi-starter
deprecation==2.1.0
    # via databricks-vectorsearch
distro==1.9.0
    # via openai
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.135.1
    # via
    #   databricks-apps-fastapi-starter
    #   mlflow-skinny
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via mlflow-skinny
google-auth==2.49.1
    # via databricks-sdk
googleapis-common-protos==1.73.0
    # via opentelemetry-exporter-otlp-proto-grpc
greenlet==3.3.2 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
grpcio==1.78.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via
    #   mlflow-skinny
    #   opentelemetry-api
jiter==0.13.0
    # via openai
lz4==4.4.5
    # via databricks-sql-connector
mlflow-skinny==3.10.1
    # via databricks-vectorsearch
numpy==2.4.3
    # via
    #   databricks-apps-fastapi-starter
    #   pandas
oauthlib==3.3.1
    # via databricks-sql-connector
openai==2.28.0
    # via databricks-apps-fastapi-starter
openpyxl==3.1.5
    # via databricks-sql-connector
opentelemetry-api==1.40.0
    # via
    #   mlflow-skinny
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-distro==0.61b0
    # via databricks-apps-fastapi-starter
opentelemetry-exporter-otlp-proto-common==1.40.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.40.0
    # via databricks-apps-fastapi-starter
opentelemetry-instrumentation==0.61b0
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-asgi==0.61b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.61b0
    # via databricks-apps-fastapi-starter
opentelemetry-instrumentation-httpx==0.61b0
    # via databricks-apps-fastapi-starter
opentelemetry-instrumentation-logging==0.61b0
    # via databricks-apps-fastapi-starter
opentelemetry-instrumentation-requests==0.61b0
    # via databricks-apps-fastapi-starter
opentelemetry-instrumentation-sqlalchemy==0.61b0
    # via databricks-apps-fastapi-starter
opentelemetry-proto==1.40.0
    # via
    #   mlflow-skinny
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.40.0
    # via
    #   mlflow-skinny
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.61b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-sdk
opentelemetry-util-http==0.61b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
packaging==26.0
    # via
    #   deprecation
    #   mlflow-skinny
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-sqlalchemy
pandas==2.3.3
    # via
    #   databricks-apps-fastapi-starter
    #   databricks-sql-connector
protobuf==5.29.6
    # via
    #   databricks-sdk
    #   databricks-vectorsearch
    #   googleapis-common-protos
    #   mlflow-skinny
    #   opentelemetry-proto
pyarrow==23.0.1
    # via databricks-apps-fastapi-starter
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pybreaker==1.4.1
    # via databricks-sql-connector
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   databricks-apps-fastapi-starter
    #   fastapi
    #   mlflow-skinny
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via databricks-apps-fastapi-starter
pyjwt==2.12.1
    # via databricks-sql-connector
python-dateutil==2.9.0.post0
    # via
    #   databricks-sql-connector
    #   pandas
python-dotenv==1.2.2
    # via
    #   databricks-apps-fastapi-starter
    #   mlflow-skinny
    #   pydantic-settings
python-multipart==0.0.22
    # via databricks-apps-fastapi-starter
pytz==2026.1.post1
    # via pandas
pyyaml==6.0.3
    # via mlflow-skinny
requests==2.32.5
    # via
    #   databricks-sdk
    #   databricks-sql-connector
    #   databricks-vectorsearch
    #   mlflow-skinny
secure==1.0.1
    # via databricks-apps-fastapi-starter
six==1.17.0
    # via
    #   python-dateutil
    #   thrift
smmap==5.0.3
    # via gitdb
sniffio==1.3.1
    # via openai
sqlalchemy==2.0.48
    # via databricks-apps-fastapi-starter
sqlparse==0.5.5
    # via mlflow-skinny
starlette==0.46.2
    # via fastapi
thrift==0.20.0
    # via databricks-sql-connector
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   grpcio
    #   mlflow-skinny
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via
    #   databricks-sql-connector
    #   requests
uvicorn==0.42.0
    # via
    #   databricks-apps-fastapi-starter
    #   mlflow-skinny
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-sqlalchemy
zipp==3.23.0
    # via importlib-metadata
